created bucket generator. To run:
import bucket_list as bl
bl.main()    #to import neccessary packages and unpickle variables
bl.generateBucket(x)    #where x is the bucket number you want to iterate

The output is a list of 36 'games', in the same format as those in Wil's games2p3s.py (so we can use those tools).
This list of 32 'games' is called a "bucket".
Bucket's have the following properties:

All buckets contain 36 'games'.
All equivalent 'games' exist in the same bucket.
All buckets are disjoint.
There are 50,808,240 buckets, numbered 0 - 50,808,239.

The games are